refactor(home hero): better loading state & data fetching
