+++ apps/frontend/src/widgets/home-hero/HeroRecipe.vue
