+++ apps/frontend/src/widgets/home-hero/index.ts
@@ -1,0 +1,1 @@
+export { default as HomeHero } from "./Hero.vue";
